增加可用avaliable types

+++ src/main/java/com/siival/telegrambots/core/webhook/WebhookInfo.java
@@ -10,7 +10,7 @@ public class WebhookInfo {
     private boolean has_custom_certificate;
     private Integer pending_update_count;
     private String ip_address;
-    private Integer last_error_date;
+    private Long last_error_date;
     private String last_error_message;
     private Integer max_connections;
     private List<String> allowed_updates;
